+++ www/admin/delete_pista.php
@@ -7,6 +7,13 @@
 </head>
 <body>
 <? 
+
+session_start();
+if(!$_SESSION['shouldPass']){
+	header("Location: http://uninorterally1.hol.es");
+    exit;
+}
+
 include('config.php'); 
 $id_pista = (int) $_GET['id_pista']; 
 mysql_query("DELETE FROM `Pista` WHERE `id_pista` = '$id_pista' ") ;
