+++ www/admin/delete_pregunta.php
@@ -1,4 +1,20 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<meta charset="utf-8">
+	<link rel="stylesheet" href="f/css/normalize.css">
+	<link rel="stylesheet" href="f/css/foundation.css">
+</head>
+<body>
+
 <? 
+
+session_start();
+if(!$_SESSION['shouldPass']){
+	header("Location: http://uninorterally1.hol.es");
+    exit;
+}
+
 include('config.php'); 
 $id_pregunta = (int) $_GET['id_pregunta']; 
 mysql_query("DELETE FROM `Pregunta` WHERE `id_pregunta` = '$id_pregunta' ") ;
